Fixed the error on delete brands or categories that have products registered

+++ controllers/admin/adminBrandsController.js
@@ -116,35 +116,49 @@ class AdminBrandsController {
     async deleteBrand(req, res) {
 
         var ok = true;
+        var msg = '';
 
         if(req.body.id != '') {
 
             let brand = new BrandsModel();
-            brand = await brand.findBrand(req.body.id);
 
-            if(brand.haveImage == true) {
+            if(await brand.brandIsEmpty(req.body.id) == true) {
 
-                let fileExist = fs.existsSync(global.PROJECT_ROOT + '/public/' + brand.brandImage);
+                brand = await brand.findBrand(req.body.id);
+
+                if(brand.haveImage == true) {
+
+                    let fileExist = fs.existsSync(global.PROJECT_ROOT + '/public/' + brand.brandImage);
                 
-                if(fileExist) {
+                    if(fileExist) {
 
-                    // delete the storaged image
-                    fs.unlinkSync(global.PROJECT_ROOT + '/public/' + brand.brandImage);
+                        // delete the storaged image
+                        fs.unlinkSync(global.PROJECT_ROOT + '/public/' + brand.brandImage);
+
+                    }
 
                 }
 
-            }
+                ok = await brand.deleteBrand(req.body.id);
+                msg = 'Successfully deleted brand'
 
-            ok = await brand.deleteBrand(req.body.id);
+            }
+            else {
+
+                ok = false;
+                msg = 'Please remove all products from this brand first';
+
+            }
 
         }
         else {
 
             ok = false;
+            msg = 'Error on request brand';
 
         }
 
-        res.send({ok: ok});
+        res.send({ok: ok, msg: msg});
 
     }
